% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{create_server}
\alias{create_server}
\title{Create the shiny server for an nlmixr object}
\usage{
create_server(fit, ...)
}
\arguments{
\item{fit}{A nlmixr object}

\item{...}{Not used}
}
\value{
A shiny server function
}
\description{
For an example, see \code{run_app()}.
}
\seealso{
\code{\link{run_app}}
}
