% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ui.R
\name{create_ui}
\alias{create_ui}
\title{Create the shiny user interface for an nlmixr object}
\usage{
create_ui(fit, ...)
}
\arguments{
\item{fit}{A nlmixr object}

\item{...}{Not used}
}
\value{
A shiny dashboard page
}
\description{
Create the shiny user interface for an nlmixr object
}
\seealso{
Other nlmixr shiny app: 
\code{\link{run_app}()}
}
\concept{nlmixr shiny app}
