% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the shiny user interface for an nlmixr object}
\usage{
run_app(fit, ...)
}
\arguments{
\item{fit}{A nlmixr object}

\item{...}{Passed to create_ui and create_server}
}
\description{
Run the shiny user interface for an nlmixr object
}
\examples{
\dontrun{
# load your nlmixr fit object as 'fit'
run_app(fit)
}
}
\seealso{
Other nlmixr shiny app: 
\code{\link{create_ui}()}
}
\concept{nlmixr shiny app}
