% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param_limits.R
\name{set_param_limits}
\alias{set_param_limits}
\title{Set limits for parameter sliders}
\usage{
set_param_limits(estimate, limit, trans, direction, multiplier = 10)
}
\arguments{
\item{estimate}{Estimated value}

\item{limit}{The limit for estimation of the value}

\item{trans}{(Not yet used) how to translate the value from the
estimate/limit scale to the linear scale}

\item{direction}{Is the limit the \code{"lower"} or \code{"upper"} limit?
(Optional, may be automatically detected if all limits are in one direction
from or the same as the estimate.)}

\item{multiplier}{How are estimates modified when the limit is infinite (see details)}
}
\value{
New limits for use in the shiny inputs
}
\description{
This function is mainly for internal use. (It is not commonly needed by users
of the package.)
}
\details{
The multiplier is used when the limit is infinite as follows:

\itemize{
\item{Negative or positive estimate}{The \code{estimate} is multiplied or
divided by the \code{multiplier} to give the limit.}
\item{Zero estimate}{The positive or negative value of \code{multiplier} is
used as the limit.}
}
}
\keyword{Internal}
